Implement job tests
