+++ src/main/resources/application-production.properties
@@ -1,0 +1,5 @@
+ctp.projectKey=${CTP_PROJECT_KEY}
+ctp.clientId=${CTP_CLIENT_ID}
+ctp.clientSecret=${CTP_CLIENT_SECRET}
+ctp.authUrl=${CTP_AUTH_URL}
+ctp.apiUrl=${CTP_API_URL}
