+++ src/test/resources/application-test.properties
@@ -1,0 +1,7 @@
+spring.batch.job.enabled = false
+
+resource.orders = https://raw.githubusercontent.com/commercetools/commercetools-sunrise-data/re-structure-source/data/orders/orders.csv
+resource.channels = https://raw.githubusercontent.com/commercetools/commercetools-sunrise-data/re-structure-source/data/channels/channels.json
+resource.channelType = https://raw.githubusercontent.com/commercetools/commercetools-sunrise-data/re-structure-source/data/channels/channel-types.json
+resource.customerType = https://raw.githubusercontent.com/commercetools/commercetools-sunrise-data/re-structure-source/data/customers/customer-types.json
+resource.orderType = https://raw.githubusercontent.com/commercetools/commercetools-sunrise-data/re-structure-source/data/orders/order-types.json
